SubscriberEntityToConfigurationMapper created
